backend/competitor/requirements.txt
fastapi
uvicorn
google-api-python-client
python-dotenv
requests
pydantic
